Conflicts:
	doc/api/windows_communication_communicate.json
	doc/tips/directui的使用.txt

+++ build/make_aria2.sh
@@ -6,6 +6,7 @@ export CC=$COMPILER_PREFIX-gcc
 export CXX=$COMPILER_PREFIX-g++
 export CPP=$COMPILER_PREFIX-cpp
 export RANLIB=$COMPILER_PREFIX-ranlib
+export AR=$COMPILER_PREFIX-ar
 
 # set up the install pathname
 INSTALL_PATH_PREFIX=$HOME/mingw32
@@ -13,11 +14,11 @@ INSTALL_PATH_PREFIX=$HOME/mingw32
 # set up the source codes pathname
 SOURCE_PREFIX=$HOME
 C_ARES_PATH=${SOURCE_PREFIX}/c-ares-1.10.0
-OPENSSL_PATH=${SOURCE_PREFIX}/openssl-1.0.1f
+OPENSSL_PATH=${SOURCE_PREFIX}/openssl-1.0.1g
 EXPAT_PATH=${SOURCE_PREFIX}/expat-2.1.0
 ZLIB_PATH=${SOURCE_PREFIX}/zlib-1.2.8
 CPPUNIT_PATH=${SOURCE_PREFIX}/cppunit-1.12.1
-ARIA2_PATH=${SOURCE_PREFIX}/aria2-1.18.3
+ARIA2_PATH=${SOURCE_PREFIX}/aria2-1.18.5
 
 # check all the source codes tar.gz
 if [ ! -f "${C_ARES_PATH}.tar.gz" ]; then
@@ -76,7 +77,17 @@ echo "compiling c_ares..."
 tar zxf ${C_ARES_PATH}.tar.gz
 cd ${C_ARES_PATH}
 ./configure --prefix=${INSTALL_PATH_PREFIX} --host=${COMPILER_PREFIX} --enable-static
+if [ $? -ne 0 ]; then
+	echo "c_ares: configure failed."
+	exit 2
+fi
+
 make;make install
+if [ $? -ne 0 ]; then
+	echo "c_ares: make failed."
+	exit 2
+fi
+
 cd ..
 rm -rf ${C_ARES_PATH}
 
@@ -84,7 +95,17 @@ echo "compiling openssl..."
 tar zxf ${OPENSSL_PATH}.tar.gz
 cd ${OPENSSL_PATH}
 ./Configure --prefix=${INSTALL_PATH_PREFIX} mingw
-make -j 4
+if [ $? -ne 0 ]; then
+	echo "openssl: configure failed."
+	exit 2
+fi
+
+make
+if [ $? -ne 0 ]; then
+	echo "openssl: make failed."
+	exit 2
+fi
+
 make install
 cd ..
 rm -rf ${OPENSSL_PATH}
@@ -93,7 +114,17 @@ echo "compiling expat..."
 tar zxf ${EXPAT_PATH}.tar.gz
 cd ${EXPAT_PATH}
 ./configure --prefix=${INSTALL_PATH_PREFIX} --host=${COMPILER_PREFIX} --enable-static
+if [ $? -ne 0 ]; then
+	echo "expat: configure failed."
+	exit 2
+fi
+
 make;make install
+if [ $? -ne 0 ]; then
+	echo "expat: make failed."
+	exit 2
+fi
+
 cd ..
 rm -rf ${EXPAT_PATH}
 
@@ -101,7 +132,17 @@ echo "compiling zlib..."
 tar zxf ${ZLIB_PATH}.tar.gz
 cd ${ZLIB_PATH}
 ./configure --prefix=${INSTALL_PATH_PREFIX} --static
+if [ $? -ne 0 ]; then
+	echo "zlib: configure failed."
+	exit 2
+fi
+
 make;make install
+if [ $? -ne 0 ]; then
+	echo "zlib: make failed."
+	exit 2
+fi
+
 cd ..
 rm -rf ${ZLIB_PATH}
 
@@ -109,7 +150,17 @@ echo "compiling cppunit..."
 tar zxf ${CPPUNIT_PATH}.tar.gz
 cd ${CPPUNIT_PATH}
 ./configure --prefix=${INSTALL_PATH_PREFIX} --host=${COMPILER_PREFIX} --enable-static
+if [ $? -ne 0 ]; then
+	echo "cppunit: configure failed."
+	exit 2
+fi
+
 make;make install
+if [ $? -ne 0 ]; then
+	echo "cppunit: make failed."
+	exit 2
+fi
+
 cd ..
 rm -rf ${CPPUNIT_PATH}
 
@@ -117,7 +168,17 @@ echo "compiling aria2..."
 tar zxf ${ARIA2_PATH}.tar.gz
 cd ${ARIA2_PATH}
 ./configure --host=${COMPILER_PREFIX} --prefix=${INSTALL_PATH_PREFIX} --enable-static --enable-libaria2 --without-included-gettext --disable-nls --with-libcares="/lib/i386-linux-gnu" --without-gnutls --with-openssl="${INSTALL_PATH_PREFIX}/include/openssl" --without-sqlite3  --without-libxml2 --with-libexpat --with-libz --without-libgcrypt  --without-libnettle --with-cppunit-prefix=${INSTALL_PATH_PREFIX} CPPFLAGS="-I${INSTALL_PATH_PREFIX}/include" LDFLAGS="-L${INSTALL_PATH_PREFIX}/lib" PKG_CONFIG_PATH="${INSTALL_PATH_PREFIX}/lib/pkgconfig"
+if [ $? -ne 0 ]; then
+	echo "aria2: configure failed."
+	exit 2
+fi
+
 make -j 4
+if [ $? -ne 0 ]; then
+	echo "aria2: make failed."
+	exit 2
+fi
+
 make install
 cd ..
 rm -rf ${ARIA2_PATH}
